:thread: Move NavBar to react_library
