+++ skeleton/frontend/src/views/BUILD.bazel
@@ -2,6 +2,12 @@ load("@aspect_bazel_lib//lib:copy_file.bzl", "copy_file")
 load("@aspect_rules_jest//jest:defs.bzl", "jest_test")
 load("@aspect_rules_ts//ts:defs.bzl", "ts_project")
 
+copy_file(
+    name = "globalsdts",
+    src = "//react_library:globalsdts",
+    out = "globals.d.ts",
+)
+
 ts_project(
     name = "ExampleView",
     srcs = [
@@ -32,17 +38,18 @@ ts_project(
 
 copy_file(
     name = "setupTests",
-    src = "//skeleton/frontend/src:setupTests.js",
+    src = "//react_library:setupTests.js",
     out = "setupTests.js",
 )
 
 jest_test(
     name = "ExampleViewTest",
-    config = "//skeleton/frontend:jestconfig",
+    config = "//react_library:jestconfig",
     data = [
         ":ExampleView",
         ":setupTests",
-        "//skeleton/frontend:testfixtures",
+        "//react_library:testfixtures",
+        "//skeleton/frontend:packagejson",
     ],
     node_modules = "//:node_modules",
     node_options = [
